# eventconnect-backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "EventConnect API",
        "description": "API complète pour la gestion d'événements avec système de réservation",
        "contact": {
            "name": "EventConnect Team",
            "email": "support@eventconnect.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Serveur de développement local"
        }
    ],
    "paths": {
        "/register": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Inscription d'un nouvel utilisateur",
                "description": "Crée un nouveau compte utilisateur et retourne un token d'authentification",
                "operationId": "registerUser",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation",
                                    "role"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Nom complet de l'utilisateur",
                                        "type": "string",
                                        "example": "Jean Dupont"
                                    },
                                    "email": {
                                        "description": "Email unique de l'utilisateur",
                                        "type": "string",
                                        "format": "email",
                                        "example": "jean.dupont@example.com"
                                    },
                                    "password": {
                                        "description": "Mot de passe (minimum 8 caractères)",
                                        "type": "string",
                                        "minLength": 8,
                                        "example": "password123"
                                    },
                                    "password_confirmation": {
                                        "description": "Confirmation du mot de passe",
                                        "type": "string",
                                        "example": "password123"
                                    },
                                    "phone": {
                                        "description": "Numéro de téléphone (optionnel)",
                                        "type": "string",
                                        "example": "0123456789"
                                    },
                                    "role": {
                                        "description": "Rôle de l'utilisateur",
                                        "type": "string",
                                        "enum": [
                                            "participant",
                                            "organisateur"
                                        ],
                                        "example": "participant"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Inscription réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Inscription réussie"
                                        },
                                        "user": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Jean Dupont"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "jean.dupont@example.com"
                                                },
                                                "role": {
                                                    "type": "string",
                                                    "example": "participant"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "token": {
                                            "description": "Token JWT pour l'authentification",
                                            "type": "string",
                                            "example": "1|abc123..."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/login": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Connexion d'un utilisateur",
                "description": "Authentifie un utilisateur et retourne un token d'authentification",
                "operationId": "loginUser",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "Email de l'utilisateur",
                                        "type": "string",
                                        "format": "email",
                                        "example": "jean.dupont@example.com"
                                    },
                                    "password": {
                                        "description": "Mot de passe de l'utilisateur",
                                        "type": "string",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Connexion réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Connexion réussie"
                                        },
                                        "user": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Jean Dupont"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "jean.dupont@example.com"
                                                },
                                                "role": {
                                                    "type": "string",
                                                    "example": "participant"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "token": {
                                            "description": "Token JWT pour l'authentification",
                                            "type": "string",
                                            "example": "1|abc123..."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/logout": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Déconnexion de l'utilisateur",
                "description": "Déconnecte l'utilisateur et invalide le token d'authentification",
                "operationId": "logoutUser",
                "responses": {
                    "200": {
                        "description": "Déconnexion réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Déconnexion réussie"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}